@@ -12,166 +12,175 @@ namespace dauw
   }
 
   // Parse a deque of tokens into an expression
-  Expr* Parser::parse()
+  std::shared_ptr<Expr> Parser::parse()
   {
-    return parse_expression();
+    // script → (expression 'newline')* 'eof'
+    std::vector<std::shared_ptr<Expr>> exprs;
+    while (!at_end())
+    {
+      exprs.push_back(parse_expression());
+      consume("newline", "Expected newline after expression");
+    }
+    return std::make_shared<ExprBlock>(exprs);
   }
 
   // Parse an expression
-  Expr* Parser::parse_expression()
+  std::shared_ptr<Expr> Parser::parse_expression()
   {
+    // expression → operation 'newline'
     return parse_operation();
   }
 
   // Parse an operation expression
-  Expr* Parser::parse_operation()
+  std::shared_ptr<Expr> Parser::parse_operation()
   {
     // operation → logic_or
     return parse_logic_or();
   }
 
   // Parse a logic or expression
-  Expr* Parser::parse_logic_or()
+  std::shared_ptr<Expr> Parser::parse_logic_or()
   {
     // logic_or → logic_and ('or' logic_and)*
     auto expr = parse_logic_and();
-    while (match("keyword_or"))
+    while (match("operator_logical_or"))
     {
-      auto op_token = previous();
+      auto op_token = current_;
       auto right = parse_logic_and();
-      expr = new BinaryExpr(expr, op_token, right);
+      expr = std::make_shared<ExprBinary>(expr, op_token, right);
     }
     return expr;
   }
 
   // Parse a logic and expression
-  Expr* Parser::parse_logic_and()
+  std::shared_ptr<Expr> Parser::parse_logic_and()
   {
     // logic_and → logic_not ('and' logic_not)*
     auto expr = parse_logic_not();
-    while (match("keyword_and"))
+    while (match("operator_logical_and"))
     {
-      auto op_token = previous();
+      auto op_token = current_;
       auto right = parse_logic_not();
-      expr = new BinaryExpr(expr, op_token, right);
+      expr = std::make_shared<ExprBinary>(expr, op_token, right);
     }
     return expr;
   }
 
   // Parse a logic not expression
-  Expr* Parser::parse_logic_not()
+  std::shared_ptr<Expr> Parser::parse_logic_not()
   {
     // logic_not → 'not' logic_not | equality
-    if (match("keyword_not"))
+    if (match("operator_logical_not"))
     {
-      auto op_token = previous();
+      auto op_token = current_;
       auto right = parse_logic_not();
-      return new UnaryExpr(op_token, right);
+      return std::make_shared<ExprUnary>(op_token, right);
     }
 
     return parse_equality();
   }
 
   // Parse an equality expression
-  Expr* Parser::parse_equality()
+  std::shared_ptr<Expr> Parser::parse_equality()
   {
+
     // equality → comparison (('==' | '<>' | '~') comparison)*
     auto expr = parse_comparison();
     while (match({"operator_equal", "operator_not_equal", "operator_match"}))
     {
-      auto op_token = previous();
+      auto op_token = current_;
       auto right = parse_comparison();
-      expr = new BinaryExpr(expr, op_token, right);
+      expr = std::make_shared<ExprBinary>(expr, op_token, right);
     }
     return expr;
   }
 
   // Parse a comparison expression
-  Expr* Parser::parse_comparison()
+  std::shared_ptr<Expr> Parser::parse_comparison()
   {
     // comparison → threeway (('<' | '<=' | '>' | '>=' | '%%') threeway)*
     auto expr = parse_threeway();
     while (match({"operator_less", "operator_less_equal", "operator_greater", "operator_greater_equal", "operator_divisible"}))
     {
-      auto op_token = previous();
+      auto op_token = current_;
       auto right = parse_threeway();
-      expr = new BinaryExpr(expr, op_token, right);
+      expr = std::make_shared<ExprBinary>(expr, op_token, right);
     }
     return expr;
   }
 
   // Parse a threeway expression
-  Expr* Parser::parse_threeway()
+  std::shared_ptr<Expr> Parser::parse_threeway()
   {
     // threeway → range ('<=>' range)*
     auto expr = parse_range();
     while (match("operator_threeway"))
     {
-      auto op_token = previous();
+      auto op_token = current_;
       auto right = parse_range();
-      expr = new BinaryExpr(expr, op_token, right);
+      expr = std::make_shared<ExprBinary>(expr, op_token, right);
     }
     return expr;
   }
 
   // Parse a range expression
-  Expr* Parser::parse_range()
+  std::shared_ptr<Expr> Parser::parse_range()
   {
     // range → term ('..' term)?
     auto expr = parse_term();
     while (match("operator_range"))
     {
-      auto op_token = previous();
+      auto op_token = current_;
       auto right = parse_term();
-      expr = new BinaryExpr(expr, op_token, right);
+      expr = std::make_shared<ExprBinary>(expr, op_token, right);
     }
     return expr;
   }
 
   // Parse a term expression
-  Expr* Parser::parse_term()
+  std::shared_ptr<Expr> Parser::parse_term()
   {
     // term → factor (('+' | '-') factor)*
     auto expr = parse_factor();
     while (match({"operator_add", "operator_subtract"}))
     {
-      auto op_token = previous();
+      auto op_token = current_;
       auto right = parse_factor();
-      expr = new BinaryExpr(expr, op_token, right);
+      expr = std::make_shared<ExprBinary>(expr, op_token, right);
     }
     return expr;
   }
 
   // Parse a factor expression
-  Expr* Parser::parse_factor()
+  std::shared_ptr<Expr> Parser::parse_factor()
   {
     // factor → unary (('*' | '/' | '//' | '%') unary)*
     auto expr = parse_unary();
     while (match({"operator_multiply", "operator_divide", "operator_floor_divide", "operator_modulo"}))
     {
-      auto op_token = previous();
+      auto op_token = current_;
       auto right = parse_unary();
-      expr = new BinaryExpr(expr, op_token, right);
+      expr = std::make_shared<ExprBinary>(expr, op_token, right);
     }
     return expr;
   }
 
   // Parse an unary expression
-  Expr* Parser::parse_unary()
+  std::shared_ptr<Expr> Parser::parse_unary()
   {
     // unary → ('-' | '#' | '$') unary | primary
     if (match({"operator_subtract", "operator_length", "operator_string"}))
     {
-      auto op_token = previous();
+      auto op_token = current_;
       auto right = parse_primary();
-      return new UnaryExpr(op_token, right);
+      return std::make_shared<ExprUnary>(op_token, right);
     }
 
     return parse_primary();
   }
 
   // Parse a primary expression
-  Expr* Parser::parse_primary()
+  std::shared_ptr<Expr> Parser::parse_primary()
   {
     // primary → atom (call_postfix | subscript_postfix | access_postfix)*
     auto expr = parse_atom();
@@ -190,11 +199,11 @@ namespace dauw
   }
 
   // Parse a call postfix
-  Expr* Parser::parse_call_postfix(Expr* callee)
+  std::shared_ptr<Expr> Parser::parse_call_postfix(std::shared_ptr<Expr> callee)
   {
     // call_postfix → '(' arguments? ')'
     // arguments → expression (',' expression)*
-    std::vector<Expr*> arguments;
+    std::vector<std::shared_ptr<Expr>> arguments;
     if (!check("parenthesis_right"))
     {
       do {
@@ -203,15 +212,15 @@ namespace dauw
     }
 
     auto end_token = consume("parenthesis_right", "Expected ')' after call arguments");
-    return new CallExpr(callee, end_token, arguments);
+    return std::make_shared<ExprCall>(callee, end_token, arguments);
   }
 
   // Parse a subscript postfix
-  Expr* Parser::parse_subscript_postfix(Expr* callee)
+  std::shared_ptr<Expr> Parser::parse_subscript_postfix(std::shared_ptr<Expr> callee)
   {
     // subscript_postfix → '[' arguments? ']'
     // arguments → expression (',' expression)*
-    std::vector<Expr*> arguments;
+    std::vector<std::shared_ptr<Expr>> arguments;
     if (!check("square_bracket_right"))
     {
       do {
@@ -220,35 +229,35 @@ namespace dauw
     }
 
     auto end_token = consume("square_bracket_right", "Expected ']' after subscript arguments");
-    return new SubscriptExpr(callee, end_token, arguments);
+    return std::make_shared<ExprSubscript>(callee, end_token, arguments);
   }
 
   // Parse a access postfix
-  Expr* Parser::parse_access_postfix(Expr* callee)
+  std::shared_ptr<Expr> Parser::parse_access_postfix(std::shared_ptr<Expr> callee)
   {
     // access_postfix → '.' IDENTIFIER
     auto name_token = consume("identifier", "Expected identifier after accessor");
-    return new GetExpr(callee, name_token);
+    return std::make_shared<ExprGet>(callee, name_token);
   }
 
   // Parse an atom expression
-  Expr* Parser::parse_atom()
+  std::shared_ptr<Expr> Parser::parse_atom()
   {
     // atom → literal | name | list | record | lambda | parenthesized
 
     // literal → 'nothing' | 'false' | 'true' | INT | REAL | STRING
     if (match("keyword_nothing"))
-      return new LiteralExpr("nothing", previous().location());
+      return std::make_shared<ExprLiteral>("nothing", current_.location());
     if (match("keyword_false"))
-      return new LiteralExpr("false", previous().location());
+      return std::make_shared<ExprLiteral>("false", current_.location());
     if (match("keyword_true"))
-      return new LiteralExpr("true", previous().location());
+      return std::make_shared<ExprLiteral>("true", current_.location());
     if (match({"int", "real", "rune", "string", "regex"}))
-      return new LiteralExpr(previous().value(), previous().location());
+      return std::make_shared<ExprLiteral>(current_.value(), current_.location());
 
     // name → IDENTIFIER
     if (match("identifier"))
-      return new NameExpr(previous());
+      return std::make_shared<ExprName>(current_);
 
     // list → '[' (list_item (',' list_item)*)? ']'
     // list_item → expression
@@ -263,46 +272,43 @@ namespace dauw
     {
       auto expr = parse_expression();
       consume("parenthesis_right", "Expected ')' after expression");
-      return new ParenthesizedExpr(expr);
+      return std::make_shared<ExprParenthesized>(expr);
     }
 
-    throw Error("Expected expression", current().location());
+    throw Error("Expected expression", next_.location());
   }
 
-  // Advance to the next token and return the previous token
+
+  // Advance to the next token and return that token
   Token Parser::advance()
   {
     if (!at_end())
-      index_ ++;
-    return previous();
-  }
+    {
+      current_ = next_;
+      next_ = tokens_.front();
+      tokens_.pop_front();
 
-  // Return the current token
-  Token Parser::current()
-  {
-    return tokens_[index_];
-  }
+      if (current_.name() == "error")
+        throw Error(current_.value(), current_.location());
+    }
 
-  // Return the previous token
-  Token Parser::previous()
-  {
-    return tokens_[index_ - 1];
+    return current_;
   }
 
   // Return if the parser has reaced the end of the tokens
   bool Parser::at_end()
   {
-    return current().name() != "eof";
+    return next_.name() == "eof";
   }
 
-  // Check if the current token has the specified name
-  bool Parser::check(std::string name)
+  // Check if the next token has the specified name
+  bool Parser::check(string_t name)
   {
-    return at_end() ? false : current().name() == name;
+    return at_end() ? false : next_.name() == name;
   }
 
-  // Check if the current token has the specified name and advance if so
-  bool Parser::match(std::string name)
+  // Check if the next token has the specified name and advance if so
+  bool Parser::match(string_t name)
   {
     if (!check(name))
       return false;
@@ -310,7 +316,9 @@ namespace dauw
     advance();
     return true;
   }
-  bool Parser::match(const std::initializer_list<std::string> names)
+
+  // Check if the next token has one of the specified names and advance if so
+  bool Parser::match(const std::initializer_list<string_t> names)
   {
     for (auto name : names)
     {
@@ -324,11 +332,12 @@ namespace dauw
     return false;
   }
 
-  Token Parser::consume(std::string name, std::string message)
+  // Consume the next token and fail if it doesn't have the specified name
+  Token Parser::consume(string_t name, string_t message)
   {
     if (check(name))
       return advance();
 
-    throw Error(message, current().location());
+    throw Error(message, next_.location());
   }
 }
