@@ -3,22 +3,22 @@
 
 namespace dauw
 {
-  using resolver = std::function<std::string(std::string)>;
+  using resolver = std::function<string_t(string_t)>;
 
 
   // Resolve a file name
-  std::string resolve_file(std::string path)
+  string_t resolve_file(string_t path)
   {
     // Create a vector of resolvers
     std::vector<resolver> resolvers;
 
     // Canonical path
-    resolvers.push_back([](std::string path)->std::string { return path; });
-    resolvers.push_back([](std::string path)->std::string { return fmt::format("{}.dauw", path); });
+    resolvers.push_back([](string_t path)->string_t { return path; });
+    resolvers.push_back([](string_t path)->string_t { return fmt::format("{}.dauw", path); });
 
     // User bin directory
-    resolvers.push_back([](std::string path)->std::string { return fmt::format("{}/.dauw/bin/{}", std::getenv("HOME"), path); });
-    resolvers.push_back([](std::string path)->std::string { return fmt::format("{}/.dauw/bin/{}.dauw", std::getenv("HOME"), path); });
+    resolvers.push_back([](string_t path)->string_t { return fmt::format("{}/.dauw/bin/{}", std::getenv("HOME"), path); });
+    resolvers.push_back([](string_t path)->string_t { return fmt::format("{}/.dauw/bin/{}.dauw", std::getenv("HOME"), path); });
 
     // Iterate over the resolvers
     for (auto resolve : resolvers)
@@ -33,21 +33,21 @@ namespace dauw
   }
 
   // Read the contents of a file
-  std::string read_file(std::string path)
+  string_t read_file(string_t path)
   {
     std::ifstream stream(path);
-  	std::string source((std::istreambuf_iterator<char>(stream)), (std::istreambuf_iterator<char>()));
+  	string_t source((std::istreambuf_iterator<char>(stream)), (std::istreambuf_iterator<char>()));
     return source;
   }
 
 
   // Search for a regular expression in a string
-  std::optional<std::smatch> regex_search(std::regex pattern, std::string& string, size_t begin, size_t end)
+  match_optional_t regex_search(regex_t pattern, string_t& string, size_t begin, size_t end)
   {
     if (end == 0)
       end = string.length();
 
-    std::smatch match;
+    match_t match;
     if (std::regex_search(string.cbegin() + begin, string.cbegin() + end, match, pattern))
       return match;
     else
@@ -55,12 +55,12 @@ namespace dauw
   }
 
   // Match a regular expression at the beginning of a string
-  std::optional<std::smatch> regex_match(std::regex pattern, std::string& string, size_t begin, size_t end)
+  match_optional_t regex_match(regex_t pattern, string_t& string, size_t begin, size_t end)
   {
     if (end == 0)
       end = string.length();
 
-    std::smatch match;
+    match_t match;
     if (std::regex_search(string.cbegin() + begin, string.cbegin() + end, match, pattern, std::regex_constants::match_continuous))
       return match;
     else
@@ -68,12 +68,12 @@ namespace dauw
   }
 
   // Match a regular expression in a whole string
-  std::optional<std::smatch> regex_fullmatch(std::regex pattern, std::string& string, size_t begin, size_t end)
+  match_optional_t regex_fullmatch(regex_t pattern, string_t& string, size_t begin, size_t end)
   {
     if (end == 0)
       end = string.length();
 
-    std::smatch match;
+    match_t match;
     if (std::regex_match(string.cbegin() + begin, string.cbegin() + end, match, pattern))
       return match;
     else
@@ -81,9 +81,9 @@ namespace dauw
   }
 
   // Split a string by the occurrences a regular expression
-  std::vector<std::string> regex_split(std::regex pattern, std::string& string)
+  std::vector<string_t> regex_split(regex_t pattern, string_t& string)
   {
-    std::vector<std::string> splits;
+    std::vector<string_t> splits;
     std::sregex_token_iterator token(string.begin(), string.end(), pattern, -1);
     std::sregex_token_iterator end;
     for (; token != end; ++token)
@@ -92,14 +92,14 @@ namespace dauw
   }
 
   // Split a string by newlines
-  std::vector<std::string> regex_lines(std::string& string)
+  std::vector<string_t> regex_lines(string_t& string)
   {
-    std::regex pattern("\\r?\\n");
+    regex_t pattern("\\r?\\n");
     return regex_split(pattern, string);
   }
 
   // Replace non-overlapping occurrences of a regular expression in a string
-  std::string regex_sub(std::regex pattern, std::string& replacement, std::string& string)
+  string_t regex_sub(regex_t pattern, string_t& replacement, string_t& string)
   {
     return std::regex_replace(string, pattern, replacement);
   }
