@@ -4,14 +4,14 @@
 namespace dauw
 {
   // Return the string representation of the Error
-  std::string Error::what()
+  string_t Error::what()
   {
     return fmt::format("{} at {}", message(), location());
   }
 
 
   // Return the string representation of the source error
-  std::string LexerError::what()
+  string_t LexerError::what()
   {
     return fmt::format("{} while lexing the source at {}", message(), location());
   }
