@@ -4,99 +4,99 @@
 namespace dauw
 {
   // Return the name of the rule
-  std::string LexerRule::name()
+  string_t LexerRule::name()
   {
     return name_;
   }
 
   // Return the pattern of the rule
-  std::regex LexerRule::pattern()
+  regex_t LexerRule::pattern()
   {
     return pattern_;
   }
 
   // Replace a match using this rule
-  std::string LexerRule::replace(std::smatch match)
+  string_t LexerRule::replace(match_t match)
   {
     return replacement_(match);
   }
 
 
   // Constructor for the lexer
-  Lexer::Lexer(std::string name)
+  Lexer::Lexer(string_t name)
     : name_(name)
   {
     // Initialize the regex patterns
-    comment_pattern_ = std::regex("(?:--[ \\t]*(.*))|(?:\"((?:[^\\\\\"]|\\\\.)*)\")");
-    whitespace_pattern_ = std::regex("[ \\t]+");
+    comment_pattern_ = regex_t("(?:--[ \\t]*(.*))|(?:\"((?:[^\\\\\"]|\\\\.)*)\")");
+    whitespace_pattern_ = regex_t("[ \\t]+");
 
     // Initialize the rules
-    rules_.push_back(LexerRule("parenthesis_left", std::regex("\\(")));
-    rules_.push_back(LexerRule("parenthesis_right", std::regex("\\)")));
-    rules_.push_back(LexerRule("square_bracket_left", std::regex("\\[")));
-    rules_.push_back(LexerRule("square_bracket_right", std::regex("\\]")));
-    rules_.push_back(LexerRule("curly_bracket_left", std::regex("\\{")));
-    rules_.push_back(LexerRule("curly_bracket_right", std::regex("\\}")));
+    rules_.push_back(LexerRule("parenthesis_left", regex_t("\\(")));
+    rules_.push_back(LexerRule("parenthesis_right", regex_t("\\)")));
+    rules_.push_back(LexerRule("square_bracket_left", regex_t("\\[")));
+    rules_.push_back(LexerRule("square_bracket_right", regex_t("\\]")));
+    rules_.push_back(LexerRule("curly_bracket_left", regex_t("\\{")));
+    rules_.push_back(LexerRule("curly_bracket_right", regex_t("\\}")));
 
-    rules_.push_back(LexerRule("symbol_colon", std::regex(":")));
-    rules_.push_back(LexerRule("symbol_comma", std::regex(",")));
-    rules_.push_back(LexerRule("symbol_dot", std::regex("\\.")));
-    rules_.push_back(LexerRule("symbol_lambda", std::regex("\\\\")));
+    rules_.push_back(LexerRule("symbol_colon", regex_t(":")));
+    rules_.push_back(LexerRule("symbol_comma", regex_t(",")));
+    rules_.push_back(LexerRule("symbol_dot", regex_t("\\.")));
+    rules_.push_back(LexerRule("symbol_lambda", regex_t("\\\\")));
 
-    rules_.push_back(LexerRule("operator_length", std::regex("#")));
-    rules_.push_back(LexerRule("operator_string", std::regex("\\$")));
-    rules_.push_back(LexerRule("operator_multiply", std::regex("\\*")));
-    rules_.push_back(LexerRule("operator_divide", std::regex("/")));
-    rules_.push_back(LexerRule("operator_floor_divide", std::regex("//")));
-    rules_.push_back(LexerRule("operator_modulo", std::regex("%")));
-    rules_.push_back(LexerRule("operator_add", std::regex("\\+")));
-    rules_.push_back(LexerRule("operator_subtract", std::regex("-")));
-    rules_.push_back(LexerRule("operator_range", std::regex("\\.\\.")));
-    rules_.push_back(LexerRule("operator_threeway", std::regex("<=>")));
-    rules_.push_back(LexerRule("operator_less", std::regex("<")));
-    rules_.push_back(LexerRule("operator_less_equal", std::regex("<=")));
-    rules_.push_back(LexerRule("operator_greater", std::regex(">")));
-    rules_.push_back(LexerRule("operator_greater_equal", std::regex(">=")));
-    rules_.push_back(LexerRule("operator_divisible", std::regex("%%")));
-    rules_.push_back(LexerRule("operator_equal", std::regex("==")));
-    rules_.push_back(LexerRule("operator_not_equal", std::regex("!=")));
-    rules_.push_back(LexerRule("operator_match", std::regex("~")));
-    rules_.push_back(LexerRule("operator_logical_not", std::regex("not")));
-    rules_.push_back(LexerRule("operator_logical_and", std::regex("and")));
-    rules_.push_back(LexerRule("operator_logical_or", std::regex("or")));
-    rules_.push_back(LexerRule("operator_assign", std::regex("=")));
+    rules_.push_back(LexerRule("operator_length", regex_t("#"), 0));
+    rules_.push_back(LexerRule("operator_string", regex_t("\\$"), 0));
+    rules_.push_back(LexerRule("operator_multiply", regex_t("\\*"), 0));
+    rules_.push_back(LexerRule("operator_divide", regex_t("/"), 0));
+    rules_.push_back(LexerRule("operator_floor_divide", regex_t("//"), 0));
+    rules_.push_back(LexerRule("operator_modulo", regex_t("%"), 0));
+    rules_.push_back(LexerRule("operator_add", regex_t("\\+"), 0));
+    rules_.push_back(LexerRule("operator_subtract", regex_t("-"), 0));
+    rules_.push_back(LexerRule("operator_range", regex_t("\\.\\."), 0));
+    rules_.push_back(LexerRule("operator_threeway", regex_t("<=>"), 0));
+    rules_.push_back(LexerRule("operator_less", regex_t("<"), 0));
+    rules_.push_back(LexerRule("operator_less_equal", regex_t("<="), 0));
+    rules_.push_back(LexerRule("operator_greater", regex_t(">"), 0));
+    rules_.push_back(LexerRule("operator_greater_equal", regex_t(">="), 0));
+    rules_.push_back(LexerRule("operator_divisible", regex_t("%%"), 0));
+    rules_.push_back(LexerRule("operator_equal", regex_t("=="), 0));
+    rules_.push_back(LexerRule("operator_not_equal", regex_t("<>"), 0));
+    rules_.push_back(LexerRule("operator_match", regex_t("~"), 0));
+    rules_.push_back(LexerRule("operator_logical_not", regex_t("not"), 0));
+    rules_.push_back(LexerRule("operator_logical_and", regex_t("and"), 0));
+    rules_.push_back(LexerRule("operator_logical_or", regex_t("or"), 0));
+    rules_.push_back(LexerRule("operator_assign", regex_t("="), 0));
 
-    rules_.push_back(LexerRule("keyword_do", std::regex("do")));
-    rules_.push_back(LexerRule("keyword_else", std::regex("else")));
-    rules_.push_back(LexerRule("keyword_false", std::regex("false")));
-    rules_.push_back(LexerRule("keyword_for", std::regex("for")));
-    rules_.push_back(LexerRule("keyword_func", std::regex("func")));
-    rules_.push_back(LexerRule("keyword_if", std::regex("if")));
-    rules_.push_back(LexerRule("keyword_in", std::regex("in")));
-    rules_.push_back(LexerRule("keyword_nothing", std::regex("nothing")));
-    rules_.push_back(LexerRule("keyword_return", std::regex("return")));
-    rules_.push_back(LexerRule("keyword_then", std::regex("then")));
-    rules_.push_back(LexerRule("keyword_true", std::regex("true")));
-    rules_.push_back(LexerRule("keyword_until", std::regex("until")));
-    rules_.push_back(LexerRule("keyword_while", std::regex("while")));
+    rules_.push_back(LexerRule("keyword_do", regex_t("do")));
+    rules_.push_back(LexerRule("keyword_else", regex_t("else")));
+    rules_.push_back(LexerRule("keyword_false", regex_t("false")));
+    rules_.push_back(LexerRule("keyword_for", regex_t("for")));
+    rules_.push_back(LexerRule("keyword_func", regex_t("func")));
+    rules_.push_back(LexerRule("keyword_if", regex_t("if")));
+    rules_.push_back(LexerRule("keyword_in", regex_t("in")));
+    rules_.push_back(LexerRule("keyword_nothing", regex_t("nothing")));
+    rules_.push_back(LexerRule("keyword_return", regex_t("return")));
+    rules_.push_back(LexerRule("keyword_then", regex_t("then")));
+    rules_.push_back(LexerRule("keyword_true", regex_t("true")));
+    rules_.push_back(LexerRule("keyword_until", regex_t("until")));
+    rules_.push_back(LexerRule("keyword_while", regex_t("while")));
 
-    rules_.push_back(LexerRule("identifier", std::regex("[A-Za-z_][A-Za-z0-9_]*|`((?:[^\\\\`]|\\\\.)+)`"), &replace_identifier_));
+    rules_.push_back(LexerRule("identifier", regex_t("[A-Za-z_][A-Za-z0-9_]*|`((?:[^\\\\`]|\\\\.)+)`"), &replace_identifier_));
 
-    rules_.push_back(LexerRule("int", std::regex("0[Xx][0-9A-Fa-f][0-9A-Fa-f_]*|-?(?:0|[1-9][0-9_]*)"), 0));
-    rules_.push_back(LexerRule("real", std::regex("-?(?:0|[1-9][0-9_]*)(?:\\.[0-9][0-9_]*(?:[Ee][+-]?(?:0|[1-9][0-9_]*))?|[Ee][+-]?(?:0|[1-9][0-9_]*))"), 0));
-    rules_.push_back(LexerRule("rune", std::regex("'((?:[^\\\\']|\\\\.)*)'"), 1));
-    rules_.push_back(LexerRule("string", std::regex("\"((?:[^\\\\\"]|\\\\.)*)\""), 1));
-    rules_.push_back(LexerRule("regex", std::regex("/((?:[^\\\\/]|\\\\.)*)/[A-Za-z]*"), 0));
+    rules_.push_back(LexerRule("int", regex_t("0[Xx][0-9A-Fa-f][0-9A-Fa-f_]*|-?(?:0|[1-9][0-9_]*)"), 0));
+    rules_.push_back(LexerRule("real", regex_t("-?(?:0|[1-9][0-9_]*)(?:\\.[0-9][0-9_]*(?:[Ee][+-]?(?:0|[1-9][0-9_]*))?|[Ee][+-]?(?:0|[1-9][0-9_]*))"), 0));
+    rules_.push_back(LexerRule("rune", regex_t("'((?:[^\\\\']|\\\\.)*)'"), 1));
+    rules_.push_back(LexerRule("string", regex_t("\"((?:[^\\\\\"]|\\\\.)*)\""), 1));
+    rules_.push_back(LexerRule("regex", regex_t("/((?:[^\\\\/]|\\\\.)*)/[A-Za-z]*"), 0));
   }
 
   // Replacement function for an identifier
-  std::string Lexer::replace_identifier_(std::smatch match)
+  string_t Lexer::replace_identifier_(match_t match)
   {
     return match.str(1) != "" ? match.str(1) : match.str(0);
   }
 
   // Convert a string into a deque of tokens
-  std::deque<Token> Lexer::tokenize(std::string source)
+  std::deque<Token> Lexer::tokenize(string_t source)
   {
     // Initialize a deque to store the tokens
     std::deque<Token> tokens;
@@ -119,14 +119,14 @@ namespace dauw
           throw LexerError("A shebang is only allowed at the first line of the source", location);
 
         // Increase the position to past the shebang
-        location.increase_line();
+        location.increase_line_();
         continue;
       }
 
       // Skip the line if it is empty or only contains whitespace
-      if (line.empty() || line.find_first_not_of(" \t\r\n") == std::string::npos)
+      if (line.empty() || line.find_first_not_of(" \t\r\n") == string_t::npos)
       {
-        location.increase_line();
+        location.increase_line_();
         continue;
       }
 
@@ -159,7 +159,7 @@ namespace dauw
         throw LexerError("The indentation does not match any outer indentation level", location);
       }
 
-      location.increase_col(indent);
+      location.increase_col_(indent);
 
       // Iterate over the line
       while (location.col() < line.length())
@@ -172,7 +172,7 @@ namespace dauw
           if (match.str(1) != "")
           {
             tokens.push_back(Token("comment", match.str(1), location, match.length()));
-            location.increase_col(match.length());
+            location.increase_col_(match.length());
             continue;
           }
         }
@@ -182,7 +182,7 @@ namespace dauw
         if (whitespace_match.has_value())
         {
           auto match = whitespace_match.value();
-          location.increase_col(match.length());
+          location.increase_col_(match.length());
           continue;
         }
 
@@ -204,28 +204,17 @@ namespace dauw
 
         // if there is more than one matched token, then sort the current matched tokens
         if (matched_tokens.size() > 1)
-        {
-          std::sort(matched_tokens.begin(), matched_tokens.end(), [](Token a, Token b)->int {
-            if (a.location().line() != b.location().line())
-              return a.location().line() - b.location().line();
-            else if (a.location().col() != b.location().col())
-              return a.location().col() - b.location().col();
-            else if (a.length() != b.length())
-              return b.length() - a.length();
-            else
-              return 0;
-          });
-        }
+          std::sort(matched_tokens.begin(), matched_tokens.end());
 
         // Add the first token and increase the position to past the token
         auto token = *matched_tokens.begin();
         tokens.push_back(token);
-        location.increase_col(token.length());
+        location.increase_col_(token.length());
       }
 
       // Add a newline at the end of the line and update the location to the next line
       tokens.push_back(Token("newline", location));
-      location.increase_line();
+      location.increase_line_();
     }
 
     // Add dedent tokens for the remaining indents
