@@ -1,192 +1,8 @@
 ﻿#include <argh.h>
 
 #include "common.h"
-#include "ast.h"
-#include "errors.h"
-#include "lexer.h"
-#include "logging.h"
-#include "parser.h"
+#include "interpreter.h"
 
-
-class Printer : public dauw::ExprVisitor
-{
-  private:
-	  int depth = 0;
-
-	public:
-		void print(dauw::Expr* expr)
-		{
-			expr->accept(*this);
-		}
-
-		void print_depth()
-		{
-			for (int i = 0; i < depth; i ++)
-			  fmt::print("  ");
-		}
-
-		void visit_literal(dauw::LiteralExpr* expr)
-		{
-      fmt::print("{}\n", expr->value());
-		}
-
-		void visit_name(dauw::NameExpr* expr)
-		{
-      fmt::print(fmt::fg(fmt::color::slate_blue), "{}\n", expr->name());
-		}
-
-		void visit_parenthesized(dauw::ParenthesizedExpr* expr)
-		{
-			fmt::print(fmt::fg(fmt::color::orange), "parenthesized\n");
-
-			print_depth();
-			fmt::print("> ");
-
-			depth ++;
-			print(expr->nested());
-			depth --;
-		}
-
-		void visit_call(dauw::CallExpr* expr)
-		{
-			fmt::print(fmt::fg(fmt::color::orange), "call\n");
-
-			print_depth();
-			fmt::print("> ");
-
-			depth ++;
-			print(expr->callee());
-			depth --;
-
-			for (auto argument : expr->arguments())
-			{
-				print_depth();
-				fmt::print("> ");
-
-				depth ++;
-				print(argument);
-				depth --;
-			}
-		}
-
-		void visit_subscript(dauw::SubscriptExpr* expr)
-		{
-			fmt::print(fmt::fg(fmt::color::orange), "subscript\n");
-
-			print_depth();
-			fmt::print("> ");
-
-			depth ++;
-			print(expr->callee());
-			depth --;
-
-      for (auto argument : expr->arguments())
-			{
-			  print_depth();
-			  fmt::print("> ");
-
-			  depth ++;
-			  print(argument);
-			  depth --;
-		  }
-		}
-
-		void visit_get(dauw::GetExpr* expr)
-	  {
-      fmt::print(fmt::fg(fmt::color::orange), "get '{}'\n", expr->name().value());
-
-			print_depth();
-			fmt::print("> ");
-
-			depth ++;
-			print(expr->callee());
-			depth --;
-		}
-
-		void visit_unary(dauw::UnaryExpr* expr)
-		{
-			fmt::print(fmt::fg(fmt::color::orange), "unary {}\n", expr->op().value());
-
-			print_depth();
-			fmt::print("> ");
-
-			depth ++;
-			print(expr->right());
-			depth --;
-		}
-
-		void visit_binary(dauw::BinaryExpr* expr)
-		{
-			fmt::print(fmt::fg(fmt::color::orange), "binary {}\n", expr->op().value());
-
-			print_depth();
-			fmt::print("left: ");
-
-			depth ++;
-			print(expr->left());
-			depth --;
-
-			print_depth();
-			fmt::print("right: ");
-
-			depth ++;
-			print(expr->right());
-			depth --;
-		}
-};
-
-// Run a string of source code
-void run(std::string name, std::string source)
-{
-	try
-	{
-		// Convert the source string to a deque of tokens
-		dauw::Lexer lexer(name);
-		std::deque<dauw::Token> tokens = lexer.tokenize(source);
-
-    // Convert the deque of tokens to an expression
-		dauw::Parser parser(tokens);
-    dauw::Expr* expr = parser.parse();
-
-    // Print the expression
-		Printer printer;
-		printer.print(expr);
-
-		delete expr;
-	}
-	catch (dauw::Error& error)
-	{
-		dauw::log_error(error, source);
-	}
-}
-
-// Run the read-eval-print loop
-void run_prompt()
-{
-	// Variable to hold the current line
-	std::string source;
-
-	fmt::print("Use CTRL+C to exit the interactive prompt\n");
-
-	// Execute the read-eval-print loop
-	while (true)
-	{
-		// Read a new line and run it
-		fmt::print(fmt::fg(fmt::color::slate_blue ), "$ ");
-
-		std::getline(std::cin, source);
-		run("<prompt>", source);
-	}
-}
-
-// Run the file specified by the file path
-void run_file(std::string path)
-{
-	// Read and run the contents of the file
-	auto absolute_path = dauw::resolve_file(path);
-	auto source = dauw::read_file(absolute_path);
-	run(absolute_path, source);
-}
 
 // Print the header
 void print_header(int argc, const char* argv[])
@@ -216,6 +32,9 @@ void print_usage(int argc, const char* argv[])
 // Main function
 int main(int argc, const char* argv[])
 {
+  // Create the interpreter
+	auto interpreter = std::make_shared<dauw::Interpreter>();
+
 	// Parse the arguments
 	argh::parser cmdl;
 	cmdl.add_params({"-h", "--help"});
@@ -239,14 +58,14 @@ int main(int argc, const char* argv[])
 	else if (cmdl(1))
 	{
 		// Evaluate the source code in the specified file
-		run_file(cmdl(1).str());
-		return 0;
+		return interpreter->interpret_from_file(cmdl(1).str());
 	}
 	else
 	{
 		// If no file has been specified, then run the interactive prompt
 		print_header(argc, argv);
-		run_prompt();
-	  return 0;
+		fmt::print("Use CTRL+C to exit the interactive prompt\n");
+
+		return interpreter->interpret_from_repl();
 	}
 }
