@@ -5,12 +5,15 @@
 
 namespace dauw
 {
+  // Forward declarations
+  class Lexer;
+
   // Class that defines a location in a source string
   class Location
   {
     private:
       // The name of the source of the location
-      std::string name_;
+      string_t name_;
 
       // The line of the location
       size_t line_;
@@ -19,30 +22,39 @@ namespace dauw
       size_t col_;
 
 
+      // Increase the line of the location
+      void increase_line_(size_t n = 1);
+
+      // Increase the column of the location
+      void increase_col_(size_t n = 1);
+
+
     public:
       // Constructor
-      Location(std::string name, size_t line, size_t col)
+      Location(string_t name, size_t line, size_t col)
         : name_(name), line_(line), col_(col) { }
-      Location(std::string name)
+      Location(string_t name)
         : Location(name, 0, 0) { }
 
       // Return the name of the source of the location
-      std::string name();
+      string_t& name();
 
       // Return the line of the location
-      size_t line();
+      size_t& line();
 
       // Return the column of the location
-      size_t col();
-
-      // Increase the line of the location
-      void increase_line(size_t n = 1);
-
-      // Increase the column of the location
-      void increase_col(size_t n = 1);
+      size_t& col();
 
       // Return the string representation of this location in a source string
-      std::string format(std::string source);
+      string_t format(string_t& source);
+
+      // Operator overloads
+      bool operator==(const Location& other);
+      std::strong_ordering operator<=>(const Location& other);
+
+
+      // Friend classes
+      friend class Lexer;
   };
 
 
@@ -51,10 +63,10 @@ namespace dauw
   {
     private:
       // The name of the token
-      std::string name_;
+      string_t name_;
 
       // The value of the token
-      std::string value_;
+      string_t value_;
 
       // The location of the token
       Location location_;
@@ -65,22 +77,28 @@ namespace dauw
 
     public:
       // Constructor
-      Token(std::string name, std::string value, Location location, size_t length = 0)
+      Token(string_t name, string_t value, Location location, size_t length = 0)
         : name_(name), value_(value), location_(location), length_(length) { }
-      Token(std::string name, Location location, size_t length = 0)
+      Token(string_t name, Location location, size_t length = 0)
         : Token(name, "", location, length) { }
+      Token()
+        : Token("bof", "", Location("")) { }
 
       // Return the name of the token
-      std::string name();
+      string_t& name();
 
       // Return the value of the token
-      std::string value();
+      string_t& value();
 
       // Return the location of the token
-      Location location();
+      Location& location();
 
       // Return the length of the token
-      size_t length();
+      size_t& length();
+
+      // Operator overloads
+      bool operator==(const Token& other);
+      std::strong_ordering operator<=>(const Token& other);
   };
 }
 
@@ -89,30 +107,30 @@ namespace fmt
 {
   // Class that defines a formatter for a location
   template <>
-  struct formatter<dauw::Location> : formatter<std::string_view>
+  struct formatter<dauw::Location> : formatter<string_view_t>
   {
     template <typename FormatContext>
     auto format(dauw::Location location, FormatContext& ctx)
     {
-      std::string format = fmt::format("{}, line {}, col {}", location.name(), location.line() + 1, location.col() + 1);
-      return formatter<string_view>::format(format, ctx);
+      string_t format = fmt::format("{}, line {}, col {}", location.name(), location.line() + 1, location.col() + 1);
+      return formatter<string_view_t>::format(format, ctx);
     }
   };
 
 
   // Class that defines a formatter for a lexer token
   template <>
-  struct formatter<dauw::Token> : formatter<string_view>
+  struct formatter<dauw::Token> : formatter<string_view_t>
   {
     template <typename FormatContext>
     auto format(dauw::Token token, FormatContext& ctx)
     {
-      std::string format;
+      string_t format;
       if (!token.value().empty())
         format = fmt::format("{} \"{}\" at {}", token.name(), token.value(), token.location());
       else
         format = fmt::format("{} at {}", token.name(), token.location());
-      return formatter<string_view>::format(format, ctx);
+      return formatter<string_view_t>::format(format, ctx);
     }
   };
 }
